Modularized configure script.
It's now split out into functions, and the first function to execute is intuitively placed at the top of the file. Most importantly, a dedicated 'writeLine' function is used to write lines to the Makefile being generated.

Also, a check was added so that the script aborts with an error if a subproject fails to create a Makefile after its script returns.